--- basics/TODO/main.rs
@@ -1,38 +1,0 @@
-// guess game from TRLB 
-extern crate rand;
-
-use rand::Rng;
-use std::cmp::Ordering;
-use std::io;
-
-fn main() {
-    println!("Guess the number!");
-
-    let rand_number = rand::thread_rng().gen_range(1, 101);
-
-    loop {
-        println!("Please input your guess.");
-
-        let mut guess = String::new();
-
-        io::stdin()
-        .read_line(&mut guess)
-        .expect("Failed to read line");
-
-        let guess: u32 = match guess.trim().parse() {
-            Ok(num) => num,
-            Err(_) => continue,
-        };
-
-        println!("Your guess {}", guess);
-
-        match guess.cmp(&rand_number) {
-            Ordering::Less => println!("Too small"),
-            Ordering::Greater => println!("Too big"),
-            Ordering::Equal => {
-                println!("You win");
-                break;
-            }
-        }
-    }
-}
