+++ basics/TODO/time_intervals.rs
@@ -7,7 +7,7 @@
 ///```rust
 ///input = [(1,4), (2,3)]> 3    input = [(4,6), (1,2)]> 3   
 ///input = [(1,4), (6,8), (2,4), (7,9), (10, 15)]> 11
-///``` 
+///```
 
 /// 时间段类型
 type TimeIntervals = (i32, i32);
