+++ ffi/string/upper.c
