+++ ffi/string/woops.c
