+++ hands-on/httpie/.gitignore
