+++ hands-on/httpie/Cargo.toml
