+++ hands-on/httpie/src/main.rs
